% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{australian.carnivora}
\alias{australian.carnivora}
\title{Australian Carnivora dataset.}
\format{A data frame with 162034 rows and 44 variables:}
\source{
\url{https://www.gbif.org}
}
\usage{
australian.carnivora
}
\description{
A dataset containing Carnivora occurances in Australia.
}
\keyword{datasets}
