% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checks.R
\name{center_of_the_country_coordinates_flag}
\alias{center_of_the_country_coordinates_flag}
\title{Flag records with coordinates of center of the country}
\usage{
center_of_the_country_coordinates_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with two mandatory fields; "decimalLatitude", "decimalLongitude"}
}
\value{
Same dataframe with one additional column; centerofTheCountryCoordinatesFlag
}
\description{
Flag records with decimalLatitude/decimalLongitude=spatial buffered centre of country
}
\details{
When coordinates are not known auto georefences makes the center of the country the coordinate. This flag brings that out
}
\section{samplePassData}{

decimalLatitude/decimalLongitude=spatial buffered centre of country
}

\section{sampleFailData}{

decimalLatitude/decimalLongitude=spatial buffered centre of country
}

\section{targetDWCField}{

decimalLatitude, decimalLongitude
}

\section{checkCategory}{

spatial
}

\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- center_of_the_country_coordinates_flag(dat$data)

}
\author{
thiloshon <thiloshon@gmail.com>
}
