% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Data cleaning according to configuration supplied}
\usage{
<<<<<<< HEAD
clean_data(bddata, config, verbose = T, report = T,
  format = c("md_document", "html_document", "pdf_document"))
=======
clean_data(bddata, config, verbose = T, report = T)
>>>>>>> upstream/master
}
\arguments{
\item{bddata}{Biodiversity data in a data frame}

\item{config}{Configuration generated using \code{get_config}}

\item{verbose}{Verbose output if TRUE else brief output if FALSE}

<<<<<<< HEAD
\item{report}{Whether to print report of cleaning done.}

\item{format}{Formats of the cleaning report required. Options are: Markdown, HTML or / and PDF}
=======
\item{report}{Whether to print report of cleaning done. Options are: Markdown, HTML or / and PDF}
>>>>>>> upstream/master
}
\value{
data frame with clean data
}
\description{
Use \code{get_config} to generate configuration and pass it to this
function to process the data accordingly.
}
\examples{
\dontrun{
library(rgbif)
occdat1 <- occ_data(
 country = "AU",     # Country code for australia
 classKey= 359,      # Class code for mammalia
 limit=5000,         # Get only 5000 records
 )
 myData<-occdat1$data
 myConfig <- get_config()
 cleanData <- clean_data(myData,myConfig)
 cleanData <- clean_data(myData,myConfig, report = T)
}

}
