% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_new.R
\name{clean_data_new}
\alias{clean_data_new}
\title{Data cleaning according to Questionnaire Responses.}
\usage{
clean_data_new(data, customQuestionnaire = NULL)
}
\arguments{
\item{customQuestionnaire}{Custom User Created Questionnaire Responses if
already available to pypass asking questions each time.}
}
\value{
data frame with clean data
}
\description{
Use \code{run_questionnaire} to create Questionnaire Responses and pass it to this
function to process the data faster.
}
\details{
Use \code{create_default_questionnaire} to create default Questionnaire object.
You can add your custom questions and then pass it to this
function to process the data.
}
\examples{
\dontrun{
library(rgbif)
occdat1 <- occ_data(
 country = "AU",     # Country code for australia
 classKey= 359,      # Class code for mammalia
 limit=5000,         # Get only 5000 records
 )
 myData<-occdat1$data

 cleanedData <- clean_data_new(myData)

 responses <- run_questionnaire()
 cleanedData <- clean_data_new(myData, responses)

 customQuestionnaire <- create_default_questionnaire()
 customResponses <- run_questionnaire(customQuestionnaire)
 cleanedData <- clean_data_new(myData, customResponses)
 }

}
