% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_new.R
\name{cleanData}
\alias{cleanData}
\title{Data cleaning according to Questionnaire Responses.}
\usage{
cleanData(data, customQuestionnaire = NA)
}
\arguments{
\item{customQuestionnaire}{Custom User Created Questionnaire Responses if
already available to pypass asking questions each time.}
}
\value{
data frame with clean data
}
\description{
Use \code{runQuestionnaire} to create Questionnaire Responses and pass it to this
function to process the data faster.
}
\details{
Use \code{createDefaultQuestionnaire} to create default Questionnaire object.
You can add your custom questions and then pass it to this
function to process the data.
}
\examples{
\dontrun{
library(rgbif)
occdat1 <- occ_data(
 country = "AU",     # Country code for australia
 classKey= 359,      # Class code for mammalia
 limit=5000,         # Get only 5000 records
 )
 myData<-occdat1$data

 cleanedData <- cleanData(myData)

 responses <- runQuestionnaire()
 cleanedData <- cleanData(myData, responses)

 customQuestionnaire <- createDefaultQuestionnaire()
 customResponses <- runQuestionnaire(customQuestionnaire)
 cleanedData <- cleanData(myData, customResponses)
 }

}
