% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_making.R
\name{cleaning_function}
\alias{cleaning_function}
\title{Data decision function (binary decision) required in bdclean internal usage.}
\usage{
cleaning_function(bddata)
}
\arguments{
\item{bddata}{The dataframe to clean}
}
\description{
NOTE: This is an package internal function. Do not use for external uses. Exported to make it available for shiny app.
}
\examples{

if(interactive()){

library(rgbif)
occdat <- occ_data(
  country = 'AU', # Country code for australia
  classKey = 359, # Class code for mammalia
  limit = 50 # Get only 50 records
)
myData <- occdat$data
cleaned_data <- cleaning_function(myData)

} 

}
