% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checks.R
\name{coordinate_precision_outofrange_flag}
\alias{coordinate_precision_outofrange_flag}
\title{Flag records with incorrect coordinatePrecision}
\usage{
coordinate_precision_outofrange_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with one mandatory field; coordinatePrecision}
}
\value{
Same dataframe with one additional column; coordinatePrecisionOutofRangeFlag
}
\description{
Runs quality check of checking if coordinatePrecision is wthin possible range.
}
\details{
The coordinate precision (dwc:coordinatePrecision), as a decimal representation, is outside the range of zero (minimum) and
one (maximum) coordinatePrecision /=>0<=1.
coordinatePrecision is a measure of precision of the coordinates. It can take only values between 0 and 1 for 1 being
highly precise and 0 being no precise at all. This check will flag all records with precision out of this range.
}
\section{samplePassData}{

0 <= coordinatePrecision <= 1
}

\section{sampleFailData}{

0 <= coordinatePrecision <= 1
}

\section{targetDWCField}{

coordinatePrecision
}

\section{checkCategory}{

spatial
}

\examples{
dat <- finch::dwca_read("sources/GBIF Downloaded Files/DwC/0090371-160910150852091.zip", read=TRUE)
flagged_dat <- coordinate_precision_outofrange_flag(dat$data$occurrence.txt)

}
\author{
thiloshon <thiloshon@gmail.com>
}
