% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire.R
\name{create_alternate_questionnaire_1}
\alias{create_alternate_questionnaire_1}
\title{Create the alternate Questionnaire.}
\usage{
create_alternate_questionnaire_1()
}
\value{
BdQuestionContainer object with default Questions
}
\description{
Create the alternate Questionnaire.
}
\examples{

customQuestionnaire <- create_default_questionnaire()

}
