% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_questionnaire.R
\name{createDefaultQuestionnaire}
\alias{createDefaultQuestionnaire}
\title{Create the package default Questionnaire.}
\usage{
createDefaultQuestionnaire()
}
\value{
BdQuestionContainer object with default Questions
}
\description{
Create the package default Questionnaire.
}
\examples{
\dontrun{
library(rgbif)
occdat1 <- occ_data(
 country = "AU",     # Country code for australia
 classKey= 359,      # Class code for mammalia
 limit=5000,         # Get only 5000 records
 )
 myData<-occdat1$data
 
 responses <- runQuestionnaire()
 cleanedData <- cleanData(myData, responses)
 
 customQuestionnaire <- createDefaultQuestionnaire()
 customResponses <- runQuestionnaire(customQuestionnaire)
 cleanedData <- cleanData(myData, customResponses)
}

}
