% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checks.R
\name{earliestDate}
\alias{earliestDate}
\title{Clean data based on earliest date.}
\usage{
earliestDate(bddata, res = "1700-01-01")
}
\arguments{
\item{bddata}{Bio diversity data in a data frame}

\item{res}{The earliest data required}
}
\description{
Clean data based on earliest date.
}
\section{samplePassData}{

When resolution is 20-Jan-2005, records recorded after the date will pass.
}

\section{sampleFailData}{

When resolution is 20-Jan-2005, records recorded before the date will fail.
}

\section{targetDWCField}{

eventDate
}

\section{checkCategory}{

temporal
}

