% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checks.R
\name{georeference_post_occurrence_flag}
\alias{georeference_post_occurrence_flag}
\title{Flag records georeferenced after the event date}
\usage{
georeference_post_occurrence_flag(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with two mandatory fields; georeferencedDate and eventDate}
}
\value{
Same dataframe with one additional column; georeferencePostOccurrenceFlag
}
\description{
Runs quality check of checking if georeferencing was done after event date.
}
\details{
The function collects the time in years between the occurrence date and the georeferenced date.
If the record was georeferenced on the day it was occurred, then the record will most likely to be correct. But,
if the record was georeferenced several years later, then the reliability will be low. This flag brings that out.
}
\section{samplePassData}{

If the record was georeferenced on the day it was occurred
}

\section{sampleFailData}{

If the record was georeferenced several years later
}

\section{targetDWCField}{

georeferencedDate, eventDate
}

\section{checkCategory}{

temporal
}

\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- georeference_post_occurrence_flag(dat$data)

}
\author{
thiloshon <thiloshon@gmail.com>
}
