% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_making.R
\name{get_checks_list}
\alias{get_checks_list}
\title{Returning checks list, function required in bdclean internal usage.}
\usage{
get_checks_list()
}
\description{
NOTE: This is an package internal function. Do not use for external uses.
}
\examples{

if(interactive()){

all_checks <- get_checks_list()

} 

}
