% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config}
\alias{get_config}
\title{Get user inputs about data cleaning protocol to follow}
\usage{
get_config(quest = NULL)
}
\arguments{
\item{quest}{data.frame containing set of questions to be asked}
}
\value{
data frame with the responces, which would be passed on to \code{clean_data}
}
\description{
\code{get_config} asks user a set of questions and the answers are stored
in configuration variable. This variable will be used to clean the data
depending on the responces
}
\examples{
\dontrun{
myConfig <- get_config()
}

}
