% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{get_user_response}
\alias{get_user_response}
\title{Internal function for getting user response}
\usage{
get_user_response(bd_question)
}
\arguments{
\item{bd_question}{The BDQuestion object to get users responses.}
}
\description{
Internal function for getting user response
}
