% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{getUserResponse}
\alias{getUserResponse}
\title{Internal function for getting user response}
\usage{
getUserResponse(bdQuestion)
}
\arguments{
\item{bdQuestion}{The BDQuestion object to get users responses.}
}
\description{
Internal function for getting user response
}
