% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{misNames}
\alias{misNames}
\title{Data cleaning according to configuration supplied}
\usage{
misNames(bddata, res = "No")
}
\description{
Use \code{get_config} to generate configuration and pass it to this
function to process the data accordingly.
}
