% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_making.R
\name{perform_Cleaning}
\alias{perform_Cleaning}
\title{Data decision function (threshold tuning) required in bdclean internal usage.}
\usage{
perform_Cleaning(flaggedData, cleaningThreshold = 5)
}
\arguments{
\item{flaggedData}{The dataset with flags to be cleaned.}

\item{cleaningThreshold}{The Cleaning tolerance. Not used in current version.}
}
\description{
NOTE: This is an package internal function. Do not use for external uses.
}
