% Generated by roxygen2: do not edit by hand
\docType{data}
\name{quest}
\alias{quest}
\title{quest}
\description {This data set contains various questions which are displayed to the user for filtering the data based on the responses to these questions}
\format{A data frame with 6 rows and 8 columns:
\describe{
  \item{qvar}{question name}
  \item{rtype}{type of response}
  \item{rlimit}{response limit}
  \item{qcode}{question code}
  \item{qlink}{linking of question}
  \item{rescond}{}
  \item{mtype}{}
}

data set part of GSOC-2017 Project; data set under development}
\keyword{datasets}
\keyword{internal}