% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checks.R
\name{repeating_digits}
\alias{repeating_digits}
\title{Flag repeating digits}
\usage{
repeating_digits(gbif_data)
}
\arguments{
\item{gbif_data}{Dataframe from GBIF with two mandatory fields; decimalLatitude and decimalLongitude.}
}
\value{
Same dataframe with two additional columns; latRepeatCount and longRepeatCount. Both shows the number of digits that are repeating
}
\description{
Runs quality check of finding repeated digits and flags accordingly.
}
\details{
The function runs a quality check on two fields of GBIF data - decimalLatitude and decimalLongitude.
Checks if the decimal values have repeated values at the end of the value. Repeated values might
mean error in digitisation or conversions of records
}
\section{samplePassData}{

lat - 19.5518, long 113.8497.
}

\section{sampleFailData}{

lat - 19.5555, long 113.3777.
}

\section{targetDWCField}{

decimalLongitude, decimalLatitude
}

\section{checkCategory}{

spatial
}

\examples{
dat <- rgbif::occ_data(scientificName = 'Ursus americanus')
flagged_dat <- repeating_digits(dat$data)

}
\author{
thiloshon <thiloshon@gmail.com>
}
