% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{run_questionnaire}
\alias{run_questionnaire}
\title{Execute the Questionnaire and save user responses.}
\usage{
run_questionnaire(custom_questionnaire = NULL)
}
\arguments{
\item{custom_questionnaire}{Custom User Created Questionnaire if already available.}
}
\value{
list with BdQuestionObjects containing user answers
}
\description{
Execute the Questionnaire and save user responses.
}
\examples{

if(interactive()){ 

responses <- run_questionnaire()

}

}
