% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_questionnaire.R
\name{runQuestionnaire}
\alias{runQuestionnaire}
\title{Execute the Questionnaire and save user responses.}
\usage{
runQuestionnaire(customQuestionnaire = NA)
}
\arguments{
\item{customQuestionnaire}{Custom User Created Questionnaire if already available.}
}
\value{
list with BdQuestionObjects containing user answers
}
\description{
Execute the Questionnaire and save user responses.
}
\examples{
\dontrun{
library(rgbif)
occdat1 <- occ_data(
 country = "AU",     # Country code for australia
 classKey= 359,      # Class code for mammalia
 limit=5000,         # Get only 5000 records
 )
 myData<-occdat1$data
 
 responses <- runQuestionnaire()
 cleanedData <- cleanData(myData, responses)
}

}
