% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checks.R
\name{spatialResolution}
\alias{spatialResolution}
\title{Clean data based on spatial resolution}
\usage{
spatialResolution(bddata, res = 100)
}
\description{
Clean data based on spatial resolution
}
\section{samplePassData}{

When resolution is 100 meters, Coordinate Uncertainities below 100 meteres will pass.
}

\section{sampleFailData}{

When resolution is 100 meters, Coordinate Uncertainities above 100 meteres will fail.
}

\section{targetDWCField}{

coordinateUncertaintyInMeters
}

\section{checkCategory}{

spatial
}

