% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdclean.R
\docType{package}
\name{bdclean}
\alias{bdclean}
\alias{bdclean-package}
\title{bdclean: Biodiversity Data Cleaning Workflows}
\description{
Biodiversity Data Cleaning Workflows using R would be helpful
to clean biodiversity occurrence data typically downloaded from
Global Biodiversity Information Facility (http://www.gbif.org/) or
similar biodiversity data portals. There are several data cleaning
operations needed to be performed on most of the data downloaded,
in order to achieve minimum quality to use the data further for any
analysis or modelling.
}
\section{Data cleaning}{

\itemize{
\item \link{get_config}
\item \link{clean_data}
}
}

\section{Citation}{

\itemize{
\item Gueta, T., Barve, V., Nagarajah, T., Agrawal, A. & Carmel, Y. (2017). bdclean: Biodiversity data cleaning workflows (R package V 0.0.1). Retrieved from https://github.com/vijaybarve/bdclean
}
}

