% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_classes.R
\docType{class}
\name{BdQuestionContainer-class}
\alias{BdQuestionContainer-class}
\alias{BdQuestionContainer}
\title{The Question Container Reference Class}
\description{
The Question Container Reference Class
}
\section{Methods}{

\describe{
\item{\code{initialize(bdquestions = NA)}}{Construct an instance of BdQuestionContainer after validating the type.}
}}

