% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Data cleaning according to Questionnaire Responses.}
\usage{
clean_data(data, customQuestionnaire, verbose = FALSE, report = TRUE,
  format = c("html_document", "pdf_document"))
}
\arguments{
\item{data}{Biodiversity data in a data frame}

\item{customQuestionnaire}{Custom User Created Questionnaire Responses if
already available to pypass asking questions each time.}

\item{verbose}{Verbose output if TRUE else brief output if FALSE}

\item{report}{Whether to print report of cleaning done.}

\item{format}{Formats of the cleaning report required. Options are: Markdown, HTML or / and PDF}
}
\value{
data frame with clean data
}
\description{
Use \code{run_questionnaire} to create Questionnaire Responses and pass it to this
function to process the data faster.
}
\details{
Use \code{create_default_questionnaire} to create default Questionnaire object.
You can add your custom questions and then pass it to this
function to process the data.
}
\examples{
\dontrun{
library(rgbif)
occdat1 <- occ_data(
 country = "AU",     # Country code for australia
 classKey= 359,      # Class code for mammalia
 limit=5000,         # Get only 5000 records
 )
 myData<-occdat1$data

 cleanedData <- clean_data_new(myData)

 responses <- run_questionnaire()
 cleanedData <- clean_data(myData, responses)

 customQuestionnaire <- create_default_questionnaire()
 customResponses <- run_questionnaire(customQuestionnaire)
 cleanedData <- clean_data_new(myData, customResponses)
 }

}
