% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_making.R
\name{cleaning_function}
\alias{cleaning_function}
\title{Data decision function (binary decision) required in bdclean internal usage.}
\usage{
cleaning_function(bddata)
}
\arguments{
\item{bddata}{The dataframe to clean}
}
\description{
NOTE: This is an package internal function. Do not use for external uses.
}
