% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire.R
\name{create_default_questionnaire}
\alias{create_default_questionnaire}
\title{Create the package default Questionnaire.}
\usage{
create_default_questionnaire()
}
\value{
BdQuestionContainer object with default Questions
}
\description{
Create the package default Questionnaire.
}
\examples{

customQuestionnaire <- create_default_questionnaire()

}
