% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{create_report_data}
\alias{create_report_data}
\title{Generate data required to create report, function required in bdclean internal usage.}
\usage{
create_report_data(input_data, flagged_data, cleaned_data, responses,
  cleaning_true, format)
}
\arguments{
\item{input_data}{The input dataframe before cleaning}

\item{flagged_data}{The flagged data for cleaning}

\item{cleaned_data}{The data with flagged records removed}

\item{responses}{The BDQuestions object with user responses}

\item{cleaning_true}{Flag specifying if the cleaning should be done, or just flagging}

\item{format}{The format of the report to be generated}
}
\description{
NOTE: This is an package internal function. Do not use for external uses.
}
