% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{create_report_data}
\alias{create_report_data}
\title{Genrate data required to create report, function required in bdclean internal usage.}
\usage{
create_report_data(inputData, flaggedData, cleanedData, responses, cleaningTrue,
  format)
}
\description{
NOTE: This is an package internal function. Do not use for external uses.
}
