% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config}
\alias{get_config}
\title{Get user inputs about data cleaning protocol to follow}
\usage{
get_config(quest = NULL, options = NULL)
}
\arguments{
\item{quest}{data.frame containing set of questions to be asked {not implemented  yet}}

\item{options}{list of data frames with options for each questions {not implemented yet}}
}
\value{
data frame with the responses
}
\description{
\code{get_config} asks user a set of questions and the answers are stoted
in configuration variable. This variable will be used to clean the data
depending on the responses
}
\examples{
\dontrun{
myConfig <- get_config()
}

}
