% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_making.R
\name{perform_Cleaning}
\alias{perform_Cleaning}
\title{Data decision function (threshold tuning) required in bdclean internal usage.}
\usage{
perform_Cleaning(flagged_data, cleaning_threshold = 5)
}
\arguments{
\item{flagged_data}{The dataset with flags to be cleaned.}

\item{cleaning_threshold}{The Cleaning tolerance. Not used in current version.}
}
\description{
NOTE: This is an package internal function. Do not use for external uses.
}
\examples{

if(interactive()){

library(rgbif)
occdat <- occ_data(
  country = 'AU', # Country code for australia
  classKey = 359, # Class code for mammalia
  limit = 50 # Get only 50 records
)
myData <- occdat$data
cleaned_data <- perform_Cleaning(myData)

} 
}
