% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire.R
\name{run_questionnaire}
\alias{run_questionnaire}
\title{Execute the Questionnaire and save user responses.}
\usage{
run_questionnaire(customQuestionnaire = NA)
}
\arguments{
\item{customQuestionnaire}{Custom User Created Questionnaire if already available.}
}
\value{
list with BdQuestionObjects containing user answers
}
\description{
Execute the Questionnaire and save user responses.
}
\examples{
\dontrun{
library(rgbif)
occdat1 <- occ_data(
 country = "AU",     # Country code for australia
 classKey= 359,      # Class code for mammalia
 limit=5000,         # Get only 5000 records
 )
 myData<-occdat1$data
 
 responses <- run_questionnaire()
 cleanedData <- clean_data_new(myData, responses)
}

}
