% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checks.R
\name{taxoLevel}
\alias{taxoLevel}
\title{Clean data based on lower taxon level}
\usage{
taxoLevel(bddata, res = "SPECIES")
}
\arguments{
\item{bddata}{Bio diversity data in a data frame}

\item{res}{The low rank of species required}
}
\description{
Clean data based on lower taxon level
}
\section{samplePassData}{

When resolution is Species, Subspecies and Species will pass.
}

\section{sampleFailData}{

When resolution is Species, Family or Genus or any lower ranks will fail.
}

\section{targetDWCField}{

taxonRank
}

\section{checkCategory}{

taxonomic
}

