% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checks.R
\name{temporalResolution}
\alias{temporalResolution}
\title{Clean data based on temporal resolution}
\usage{
temporalResolution(bddata, res = "Day")
}
\description{
Clean data based on temporal resolution
}
